// assets/app/Auth/types/SessionSort.ts
// Code generated by tsgen. DO NOT EDIT.

export const SessionSort = {
    CreatedAt: 'createdAt',
    LastSeenAt: 'lastSeenAt',
} as const;

export type SessionSort = (typeof SessionSort)[keyof typeof SessionSort];

export const isSessionSort = (v: unknown): v is SessionSort =>
    Object.values(SessionSort).some((value) => value === v);