// assets/shared/Grid/types/SortDirection.ts
// Code generated by tsgen. DO NOT EDIT.

export const SortDirection = {
    Ascending: 'asc',
    Descending: 'desc',
} as const;

export type SortDirection = (typeof SortDirection)[keyof typeof SortDirection];

export const isSortDirection = (v: unknown): v is SortDirection =>
    Object.values(SortDirection).some((value) => value === v);